:bug: [fix] 회원가입 로직 수정
