+++ src/main/java/com/gamegoo/controller/member/RiotController.java
@@ -4,13 +4,14 @@ import com.gamegoo.apiPayload.ApiResponse;
 import com.gamegoo.dto.member.MemberRequest;
 import com.gamegoo.service.member.RiotService;
 import io.swagger.v3.oas.annotations.Operation;
-import javax.validation.Valid;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RestController;
+
+import javax.validation.Valid;
 
 @RestController
 @RequiredArgsConstructor
@@ -21,16 +22,16 @@ public class RiotController {
     private final RiotService riotService;
 
     @PostMapping("/riot")
-    @Operation(summary = "회원가입 시 riot API를 통해 소환사명을 인증하는 API", description = "API for verifying by riot API")
+    @Operation(summary = "실제 존재하는 Riot 계정인지 검증하는 API", description = "API for verifying account by riot API")
     public ApiResponse<String> VerifyRiot(
-        @RequestBody @Valid MemberRequest.verifyRiotRequestDTO verifyRiotRequestDTO) {
+            @RequestBody @Valid MemberRequest.verifyRiotRequestDTO verifyRiotRequestDTO) {
         String gameName = verifyRiotRequestDTO.getGameName();
         String tag = verifyRiotRequestDTO.getTag();
-        String email = verifyRiotRequestDTO.getEmail();
 
-        riotService.updateMemberRiotInfo(gameName, tag, email);
+        riotService.verifyRiot(gameName, tag);
 
-        return ApiResponse.onSuccess("Riot 계정이 연동되었습니다.");
+        return ApiResponse.onSuccess("해당 Riot 계정은 존재합니다.");
 
     }
+
 }
