+++ src/main/java/com/gamegoo/domain/champion/MemberChampion.java
@@ -5,6 +5,7 @@ import com.gamegoo.domain.common.BaseDateTimeEntity;
 import lombok.*;
 
 import javax.persistence.*;
+import java.util.ArrayList;
 
 @Getter
 @Entity
@@ -32,8 +33,10 @@ public class MemberChampion extends BaseDateTimeEntity {
             this.member.getMemberChampionList().remove(this);
         }
         this.member = member;
+        if (this.member.getMemberChampionList() == null) {
+            this.member.setMemberChampionList(new ArrayList<>());
+        }
         this.member.getMemberChampionList().add(this);
-
     }
 
     public void removeMember(Member member) {
